requirements.txt
birdy==0.3.2
certifi==2018.10.15
chardet==3.0.4
Django==2.0.9
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.7
jsonfield==2.0.2
oauthlib==2.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
urllib3==1.23
